Add microkelvin persistance feature and integration Change tree traversal from using `path` to using `PathWalker`
Resolves #29
