+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.1] - 2021-06-23
+
+### Added
+- Add microkelvin persistance feature and integration
+
 ## [0.8.0] - 2021-04-21
 
 ### Changed
@@ -29,7 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Change microkelvin and canonical to version 0.4
 
-[Unreleased]: https://github.com/dusk-network/nstack/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/dusk-network/nstack/compare/v0.8.1...HEAD
+[0.8.1]: https://github.com/dusk-network/nstack/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/dusk-network/nstack/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/dusk-network/nstack/compare/v0.6.1...v0.7.0
 [0.6.1]: https://github.com/dusk-network/nstack/compare/v0.6.0...v0.6.1
