+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "nstack"
-version = "0.8.0"
+version = "0.8.1"
 authors = ["Kristoffer Ström <kristoffer@dusk.network>"]
 edition = "2018"
 description = "Arity 4 stack for the kelvin merkle toolkit"
@@ -9,6 +9,9 @@ repository = "https://github.com/dusk-network/nstack"
 keywords = ["merkle", "datastructure", "stack"]
 
 [dependencies]
-microkelvin = "0.7"
+microkelvin = "0.8"
 canonical = "0.6"
 canonical_derive = "0.6"
+
+[features]
+persistance = ["microkelvin/persistance"]
